% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobarre.R
\name{cobarre}
\alias{cobarre}
\title{cobarre}
\usage{
cobarre(df = eventi, gruppo = settore)
}
\arguments{
\item{df}{il dataframe o datatable che contiene gli eventi}

\item{gruppo}{la  variabile (colonna) chge consente il raggruppamento}
}
\value{
p, un grafico ggplot a linee
}
\description{
produce un grafico a barre degli eventi per una variabile di raggruppamento,
individuata come come colonna in un dataframe
}
\examples{

cobarre(df = eventi, gruppo = settore)

}
