% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobplot.R
\name{cobplot}
\alias{cobplot}
\title{cobplot}
\usage{
cobplot(saldo = df, gruppo = genere)
}
\arguments{
\item{saldo}{la table che contiene i saldi da rappresentare. Tipicamente generata con la funzione saldi()}

\item{gruppo}{la colonna che contiene la variabile di raggruppamento (ad esempio il genere)}
}
\value{
p, un grafico ggplot a linee
}
\description{
Plot dei saldi usando ggplot2
}
\details{
usa una data.table con i saldi per reppresentare la curca dei saldi cumulati.
}
\examples{
cosaldi(eventi, inizio, fine, genere) |>
         cobplot(gruppo = genere)

}
