% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosaldi.R
\name{cosaldi}
\alias{cosaldi}
\title{cosaldi}
\usage{
cosaldi(dati = df, data_inizio = inizio, data_fine = fine, gruppo = genere)
}
\arguments{
\item{dati}{il data frame che contiene i dati}

\item{data_inizio}{la variabile che contiene la data di inizio dell'evento}

\item{data_fine}{la varibile che contiene la data di fine dell'evento}

\item{gruppo}{variabile che contiene i raggruppamementi per i quali si generano i saldi}
}
\value{
t, una data.table con avviati, cessati, saldo
e saldo_cumulato per ogni giorno e per la variabile di raggruppamento
}
\description{
genera una data table con i saldi fra avviamenti  e cessazioni
per ogni giorno in cui si verifica almeno un evento.
accetta i parametri non quotati (senza le virgolette)
}
\details{
La funzione gestisce anche i valori NA nelle date.
In questo caso il totale degli eventi viene modificato sottraendo
i casi con date NA dal conto di avvimenti e cessazioni.
Allo stesso modo dalla tavola dei saldi vengono rimosse le righe
che sono NA per le varibili di classificazione.
}
\examples{
dsa <- cosaldi(eventi, inizio, fine, genere)

}
