% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eventi}
\alias{eventi}
\title{Eventi - eventi simulati}
\format{
Un data frame con 1000 righe and 5 variabili:
\describe{
  \item{cod}{codice della riga da 1 a 1000}
  \item{inizio}{data, inizio dell'evento}
  \item{fine}{data, fine dell'evento}
  \item{genere}{fattore con due livelli "M" e "F"}
  \item{settore}{fattore con due livelli "industria" e "servizi"}

}
}
\usage{
eventi
}
\description{
un dataset che contiene  eventi simulati con un inizio, una fine e due fattori
}
\keyword{datasets}
