% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saldi.R
\name{saldi}
\alias{saldi}
\title{Saldi}
\usage{
saldi(
  dati = df,
  data_inizio = inizio,
  data_fine = fine,
  classifica = classificatore
)
}
\arguments{
\item{dati}{il data frame che contiene i dati}

\item{data_inizio}{la variabile che contiene la data di inizio dell'evento}

\item{data_fine}{la varibile che contiene la data di fine dell'evento}

\item{classifica}{variabile che contiene la classificazione per la quale si generano i saldi}
}
\value{

}
\description{
genera una data table con i saldi fra avviamenti  e cessazioni per ogni giorno in cui si verifica almeno un evento
}
