% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coplot.R
\name{coplot}
\alias{coplot}
\title{coplot}
\usage{
coplot(saldo = df, classificazione = classif)
}
\arguments{
\item{saldo}{la table che contiene i saldi da rappresentare. Tipicamente generata con la funzione saldi()}

\item{classificazione}{la colonna che contiene la variabile di classificazione (ad esempio il genere)}
}
\value{

}
\description{
Plot dei saldi usando ggplot2
}
\details{
usa una data.table con i saldi per reppresentare la curca dei saldi cumulati.
}
\examples{
saldi(eventi, "inizio", "fine", "genere") |>
         coplot(classificazione = genere)

}
